+++ src/data_source/extension.rs
@@ -354,38 +354,38 @@ where
 {
     async fn get_block_detail(
         &self,
-        request: explorer::data_source::BlockIdentifier<Types>,
-    ) -> Result<explorer::data_source::BlockDetail<Types>, explorer::data_source::GetBlockDetailError>
+        request: explorer::query_data::BlockIdentifier<Types>,
+    ) -> Result<explorer::query_data::BlockDetail<Types>, explorer::query_data::GetBlockDetailError>
     {
         self.data_source.get_block_detail(request).await
     }
 
     async fn get_block_summaries(
         &self,
-        request: explorer::data_source::GetBlockSummariesRequest<Types>,
+        request: explorer::query_data::GetBlockSummariesRequest<Types>,
     ) -> Result<
-        Vec<explorer::data_source::BlockSummary<Types>>,
-        explorer::data_source::GetBlockSummariesError,
+        Vec<explorer::query_data::BlockSummary<Types>>,
+        explorer::query_data::GetBlockSummariesError,
     > {
         self.data_source.get_block_summaries(request).await
     }
 
     async fn get_transaction_detail(
         &self,
-        request: explorer::data_source::TransactionIdentifier<Types>,
+        request: explorer::query_data::TransactionIdentifier<Types>,
     ) -> Result<
-        explorer::data_source::TransactionDetailResponse<Types>,
-        explorer::data_source::GetTransactionDetailError,
+        explorer::query_data::TransactionDetailResponse<Types>,
+        explorer::query_data::GetTransactionDetailError,
     > {
         self.data_source.get_transaction_detail(request).await
     }
 
     async fn get_transaction_summaries(
         &self,
-        request: explorer::data_source::GetTransactionSummariesRequest<Types>,
+        request: explorer::query_data::GetTransactionSummariesRequest<Types>,
     ) -> Result<
-        Vec<explorer::data_source::TransactionSummary<Types>>,
-        explorer::data_source::GetTransactionSummariesError,
+        Vec<explorer::query_data::TransactionSummary<Types>>,
+        explorer::query_data::GetTransactionSummariesError,
     > {
         self.data_source.get_transaction_summaries(request).await
     }
@@ -393,8 +393,8 @@ where
     async fn get_explorer_summary(
         &self,
     ) -> Result<
-        explorer::data_source::ExplorerSummary<Types>,
-        explorer::data_source::GetExplorerSummaryError,
+        explorer::query_data::ExplorerSummary<Types>,
+        explorer::query_data::GetExplorerSummaryError,
     > {
         self.data_source.get_explorer_summary().await
     }
@@ -403,8 +403,8 @@ where
         &self,
         query: String,
     ) -> Result<
-        explorer::data_source::SearchResult<Types>,
-        explorer::data_source::GetSearchResultsError,
+        explorer::query_data::SearchResult<Types>,
+        explorer::query_data::GetSearchResultsError,
     > {
         self.data_source.get_search_results(query).await
     }
